fix: remove platform-specific uuid code to prevent bundler errors
The Web Crypto API is available in all browsers, Node.js, and Deno. This PR assumes that covers the majority of the environments where this polyfill will be used. A simple semi-random v4 formatted string is provided for all other cases. These changes reduce the code and complexity, and should resolve bundler errors stemming from the platform-specific imports and top-level awaits that have now been removed.