fix: remove serialization reliance on top level await and external polyfills/helpers
This PR reworks the serialization code to use JSON.stringify and JSON.parse instead of structured clone and a third-party serialization polyfill/helper. A new function to configure serialization has been added to provide developers with a way to plug back in the more advanced serialization approach if they want/need it for their app.